renamed NewExample to ExampleTemplate and added comments added file comments to other examples renamed stripe_webhook_handler to ThinEventWebhookHandler for clarity
