+++ examples/README.md
@@ -3,9 +3,14 @@
 From the examples folder, run:
 `php your_example.php`
 
+e.g.
+
+`php ThinEventWebhookHandler.php`
+
 ## Adding a new example
 
-1. Clone new_example.php
+1. Clone ExampleTemplate.php
 2. Implement your example
-3. Run it (as per above)
-4. 👍
+3. Fill out the comment at the top with a description and key steps that are being demonstrated
+4. Run it (as per above)
+5. 👍
